refactor: apply quick fixes
- Extract toMaxResCoverUrl utility to @/utils/player, deduplicating
  the YouTube thumbnail maxres regex across Player and useDominantColor
- Remove unnecessary async from cookies.ts (localStorage is sync)
- Remove loading state from CookiesModal (localStorage reads are instant)
- Deduplicate auth check in admin cookies API route via assertAdmin helper
