+++ src/app/player/page.tsx
@@ -1,26 +1,12 @@
 import { Metadata } from "next";
-import InitialPlayer from "@/components/InitialPlayer";
+import { Player } from "@/components/Player";
 import { isDirectMediaURL, isTikTokURL, isYoutubeURL } from "@/utils";
 
 type SearchParams = { [key: string]: string | string[] | undefined };
 
-export async function generateMetadata({
-  searchParams,
-}: {
-  searchParams: Promise<SearchParams>;
-}): Promise<Metadata> {
-  const params = await searchParams;
-  const url = params.url as string;
-
-  if (!url) return { title: "Moonlit Player" };
-  if (isDirectMediaURL(url)) {
-    return { title: "Local File - Moonlit" };
-  }
-  if (isTikTokURL(url)) return { title: "TikTok Video - Moonlit" };
-  if (isYoutubeURL(url)) return { title: "YouTube Video - Moonlit" };
-
-  return { title: "Moonlit Player" };
-}
+export const metadata: Metadata = {
+  title: "Moonlit Player",
+};
 
 export default async function Page({
   searchParams,
@@ -30,21 +16,10 @@ export default async function Page({
   const params = await searchParams;
   const url = params.url as string;
 
-  if (!url) return <InitialPlayer />;
+  if (!url) return <Player />;
 
-  if (isDirectMediaURL(url)) {
-    const pathname = url.startsWith("/") ? url : new URL(url).pathname;
-    const name = pathname.split("/").pop() || "";
-    const title =
-      decodeURIComponent(name).replace(/\.(mp3|m4a|mp4|webm|ogg|wav)$/i, "") ||
-      "Direct media";
-    return (
-      <InitialPlayer url={url} metadata={{ title, author: "Unknown", coverUrl: "" }} />
-    );
-  }
-
-  if (isYoutubeURL(url) || isTikTokURL(url)) {
-    return <InitialPlayer url={url} metadata={{}} />;
+  if (isDirectMediaURL(url) || isYoutubeURL(url) || isTikTokURL(url)) {
+    return <Player url={url} />;
   }
 
   throw new Error("Invalid URL provided");
